Refactor count validator related classes, specially AtLeast class
This commit includes some changes that refactor the count validator related classes to make the code more understandable
and easier to maintain. It renames the only parameter in CountValidatorAbstract::validate() to give him a more meaningful
name. It also refactor the implementation of that abstract method in the AtLeast class, applying some techniques such as
"extract method" and "guard clauses".